1. Corrección campos id de int/Integer a Long. 2. Modificación de campos de artistas para reflejar la base de datos. 3. Sustitución de métodos constructores y toString corrigiendo el número de atributos 4. Corrección del error de Categorias respecto al joinColumn de idMadre

+++ src/main/java/com/listeny/listeny/models/Notificaciones.java
@@ -18,7 +18,7 @@ public class Notificaciones implements Serializable {
     @Id
     @Column (name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "descripcion", length = 200)
     private String descripcion;
@@ -30,7 +30,7 @@ public class Notificaciones implements Serializable {
 
     }
 
-    public Notificaciones(Integer id, String descripcion, String titulo) {
+    public Notificaciones(Long id, String descripcion, String titulo) {
         this.id = id;
         this.descripcion = descripcion;
         this.titulo = titulo;
