+++ src/main/java/com/listeny/listeny/repository/CancionesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CancionesRepository extends JpaRepository<Canciones, Integer> {
+public interface CancionesRepository extends JpaRepository<Canciones, Long> {
 
 }
