+++ src/main/java/com/listeny/listeny/repository/ListasRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ListasRepository extends JpaRepository<Listas, Integer> {
+public interface ListasRepository extends JpaRepository<Listas, Long> {
 
 }
